Implement Block Style differences
